# openapi.yaml
openapi: 3.1.0
info:
  title: Collaborative Travel Planner API
  description: |-
    This is the API for the Collaborative Travel Planner application. It allows users to create, manage, and share travel itineraries. The API supports real-time collaboration features through WebSockets.
    Authentication is handled via JWT Bearer tokens obtained from Google OAuth.
  version: 1.0.0
servers:
  - url: /api
paths:
  /trips:
    get:
      summary: Get all trips for the user
      description: Retrieves a list of all trips associated with the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of trips.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
    post:
      summary: Create a new trip
      description: Creates a new trip with the provided details.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTrip'
      responses:
        '201':
          description: Trip created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
...
  /users:
    get:
      summary: Search users
      description: Search for users by name or email (for collaboration features).
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Search term to match against user names or emails
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of users matching the search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicUser'
...
components:
  schemas:
...
    Trip:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        destination:
          type: string
        createdBy:
          type: string
          format: uuid
        status:
          type: string
          enum: [planning, active, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NewTrip:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
...
    PublicUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT